Only softclip once instead of three times. Needs SSE4.1 for performance increase

+++ src/sidplayfp/residfp/resample/SincResampler.cpp
@@ -44,7 +44,7 @@ namespace reSIDfp
 
 int SincResampler::fir ( int subcycle )
 {
-	auto convolve = [] ( const int16_t* __restrict__ a, const int16_t* __restrict__ b, int bLength )
+	auto convolve = [] ( const int32_t* __restrict__ a, const int16_t* __restrict__ b, int bLength )
 	{
 		auto    out = 0;
